# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API",
        "version": "1.0"
    },
    "paths": {
...
        "/dashboard/inventories": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Récupère tous les états des lieux associés au livret de l'utilisateur.",
                "description": "Récupère tous les états des lieux associés au livret de l'utilisateur.",
                "operationId": "1a093b6ca626d23b13d0a9c88ef1e692",
                "responses": {
                    "200": {
                        "description": "Liste des états des lieux."
                    },
                    "404": {
                        "description": "Aucun inventaire trouvé."
                    }
                }
            }
        },
        "/dashboard/inventory": {
            "post": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Ajoute un nouvel état des lieux.",
                "description": "Ajoute un nouvel état des lieux.",
                "operationId": "f8282855b949d9f9dd4a899b958b0c2d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "start_date",
                                    "end_date",
                                    "client_name",
                                    "status"
                                ],
                                "properties": {
                                    "start_date": {
                                        "description": "Date de début de l'état des lieux.",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "description": "Date de fin de l'état des lieux.",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "client_name": {
                                        "description": "Nom du client.",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Statut de l'état des lieux.",
                                        "type": "string"
                                    },
                                    "client_comment": {
                                        "description": "Commentaire du client.",
                                        "type": "string"
                                    },
                                    "attachment_names": {
                                        "description": "Fichiers joints.",
                                        "type": "array",
                                        "items": {
                                            "description": "Fichiers joints.",
                                            "type": "string",
                                            "format": "file"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "L'état des lieux a été ajouté avec succès."
                    }
                }
            }
        },
        "/dashboard/inventory/status": {
            "post": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Met à jour le statut d'un état des lieux.",
                "description": "Met à jour le statut d'un état des lieux.",
                "operationId": "ee659238bac8fa612ee2f706b616d210",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status",
                                    "inventory_id"
                                ],
                                "properties": {
                                    "status": {
                                        "description": "Nouveau statut de l'état des lieux.",
                                        "type": "string"
                                    },
                                    "inventory_id": {
                                        "description": "ID de l'état des lieux.",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Le statut de l'état des lieux a été mis à jour avec succès."
                    }
                }
            }
        },
        "/dashboard/inventory/{id}": {
            "delete": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Supprime un état des lieux et ses fichiers joints.",
                "description": "Supprime un état des lieux et ses fichiers joints.",
                "operationId": "5601b8aed8afdbe835996a5d0c8b532e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "L'état des lieux a été supprimé avec succès."
                    },
                    "404": {
                        "description": "Inventaire introuvable."
                    }
                }
            }
        },
        "/dashboard/inventories/search": {
            "post": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Recherche des états des lieux selon les critères donnés.",
                "description": "Recherche des états des lieux selon les critères donnés.",
                "operationId": "03dbf1a1675c1983d85c48785155b1f4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_name",
                                    "start_date",
                                    "end_date",
                                    "status"
                                ],
                                "properties": {
                                    "client_name": {
                                        "description": "Nom du client.",
                                        "type": "string"
                                    },
                                    "start_date": {
                                        "description": "Date de début de l'état des lieux.",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "description": "Date de fin de l'état des lieux.",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "description": "Statut de l'état des lieux.",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Liste des états des lieux correspondants."
                    }
                }
            }
        },
...
    },
...
}